+++ core/header_config/tier_two_nav_items.py
@@ -8,6 +8,7 @@ WHY_INVEST_IN_THE_UK_URL = urls.international.HOME / 'content/investment/why-inv
 REGIONS_URL = urls.international.HOME / 'content/investment/regions/'
 SECTORS_URL = urls.international.HOME / 'content/investment/sectors/'
 INVESTMENT_OPPORTUNITIES_URL = urls.international.HOME / 'investment/opportunities/'
+ATLAS_HOW_WE_CAN_HELP_URL = urls.international.HOME / 'content/investment/how-we-can-help-you/'
 
 
 OVERVIEW_ABOUT = NavItem(
@@ -155,4 +156,10 @@ INVESTMENT_OPPORTUNITIES = NavItem(
     name='investment-opportunities',
     title=_('Investment Opportunities'),
     url=INVESTMENT_OPPORTUNITIES_URL
+)
+
+HOW_WE_CAN_HELP_INVESTMENT_ATLAS = NavItem(
+    name='how-we-can-help',
+    title=_('How we can help'),
+    url=ATLAS_HOW_WE_CAN_HELP_URL
 )